Add tracing to exception handling.
