+++ auth-api/src/auth_api/exceptions/__init__.py
@@ -6,8 +6,14 @@ UserException - error, status_code - user rules error
 error - a description of the error {code / description: classname / full text}
 status_code - where possible use HTTP Error Codes
 """
-import functools
+
+import traceback
+from functools import wraps
+from flask import g, jsonify
+
 from auth_api.exceptions.errors import Error
+
+from sbc_common_components.tracing.exception_tracing import ExceptionTracing
 
 
 class BusinessException(Exception):
@@ -22,4 +28,51 @@ class BusinessException(Exception):
         self.code = error.name
         self.status = error.status
         self.detail = exception
+
         # to do: log/tracing exception
+        ExceptionTracing.trace(self, traceback.format_exc())
+
+
+class UserException(Exception):
+    """Exception that adds error code and error name, that can be used for i18n support."""
+
+    def __init__(self, error, status_code, trace_back, *args, **kwargs):
+        """Return a valid UserException."""
+        super(UserException, self).__init__(*args, **kwargs)
+        self.error = error
+        self.status_code = status_code
+        self.trace_back = trace_back
+
+
+def catch_custom_exception(func):
+    """TODO just a demo function"""
+
+    @wraps(func)
+    def decorated_function(*args, **kwargs):
+        try:
+            return func(*args, **kwargs)
+        except Exception as e:
+            trace_back = traceback.format_exc()
+
+            ExceptionTracing.trace(e, trace_back)
+
+            raise UserException(e.with_traceback(None), 403, trace_back)
+
+    return decorated_function
+
+
+def handle_auth_error(exception):
+    """TODO just a demo function"""
+    return jsonify(exception), exception.status_code
+
+
+def handle_db_exception(error):
+    """TODO just a demo function"""
+    return {'message': str(error.error)}, error.status_code
+
+
+def handle_exception(exception):
+    """TODO just a demo function"""
+    return {'message': str(exception.error)}, exception.status_code
+
+
