+++ auth-api/src/auth_api/resources/token.py
@@ -22,7 +22,6 @@ from auth_api import status as http_status
 from auth_api import tracing as _tracing
 from auth_api.services.keycloak import KeycloakService
 from auth_api.utils.util import cors_preflight
-from auth_api.exceptions import BusinessException
 
 from sbc_common_components.tracing.trace_tags import TraceTags
